--eslint and prettier added. --block functions refactored into blockHandler module
