+++ README.md
@@ -3,4 +3,4 @@
 This is a short script that will read through your followers and block anyone who is following greater than the number 
 you set in the .env file. 
 
-You can schedule the script by running `node scheduler.js`. Currently it is set to execute every hour.
+You can schedule the script by running `node scheduler.js`. Currently, it is set to execute every hour.
